fix(core): Support nested default content [canary] (#1964)
## Changes
- Ensure that nested calls to `withDefaultContent` work as expected

## Technical Details
`DefaultContentNode` previously inherited the store getter from the original `DefaultContentNode` and passed it on to any child `DefaultContentNodes` -- this meant that the default content it provided would not appear as existing content for an inner `DefaultContentNode`.  The fix uses  a peer of the the parent node to get the store content.

## Test Instructions
Wrap a component in nested `withDefaultContent` HOCs, or (more realistically) wrap one component in ``withDefaultContent` and render it in the context of another component also so wrapped.

This case is available on /api/default-content.